+++ src/action.py
@@ -17,17 +17,33 @@ FAIL_NAVLINK_LINK = "<not created due to error>"
 NOT_DELETE_REASON = "delete_topics is false"
 
 
-def _absolute_url(url: types_.Url | None, discourse: Discourse) -> types_.Url | None:
+def _absolute_url(
+    action: types_.AnyAction, url: types_.Url | None, discourse: Discourse
+) -> types_.Url | None:
     """Get the absolute URL.
 
     Args:
+        action: The associated action for retrieving the URL.
         url: The url to convert.
         discourse: A client to the documentation server.
 
     Returns:
         The converted url or None if the url is None.
     """
-    return discourse.absolute_url(url=url) if url is not None else None
+    if (
+        isinstance(
+            action,
+            (
+                types_.CreatePageAction,
+                types_.UpdatePageAction,
+                types_.NoopPageAction,
+                types_.DeletePageAction,
+            ),
+        )
+        and url is not None
+    ):
+        return discourse.absolute_url(url=url)
+    return url
 
 
 def _log_content_change(base: str, new: str) -> None:
@@ -56,12 +72,19 @@ def _create(
 
     Returns:
         A report on the outcome of executing the action.
+
+    Raises:
+        NotImplementedError: if a requested action has not been implemented yet.
     """
     logging.info("dry run: %s, action: %s", dry_run, action)
 
     # Handle the directory/ group case, no server interactions are required
-    if action.content is None:
+    if isinstance(action, types_.CreateGroupAction):
         url = None
+        result = types_.ActionResult.SKIP if dry_run else types_.ActionResult.SUCCESS
+        reason = DRY_RUN_REASON if dry_run else None
+    elif isinstance(action, types_.CreateExternalRefAction):
+        url = action.navlink_value
         result = types_.ActionResult.SKIP if dry_run else types_.ActionResult.SUCCESS
         reason = DRY_RUN_REASON if dry_run else None
     # Handle the file/ page case when dry run is enabled, no server interactions are required
@@ -70,7 +93,7 @@ def _create(
         result = types_.ActionResult.SKIP
         reason = DRY_RUN_REASON
     # Handle the file/ page case where a new page needs to be created on the server
-    else:
+    elif isinstance(action, types_.CreatePageAction):
         try:
             url = discourse.create_topic(
                 title=f"{name} docs: {action.navlink_title}", content=action.content
@@ -81,6 +104,9 @@ def _create(
             url = FAIL_NAVLINK_LINK
             result = types_.ActionResult.FAIL
             reason = str(exc)
+    # This should not occur since all the create action types have been implemented
+    else:  # pragma: no cover
+        raise NotImplementedError(f"internal error: {action} has not been implemented")
 
     table_row = types_.TableRow(
         level=action.level,
@@ -105,7 +131,7 @@ def _noop(action: types_.NoopAction, discourse: Discourse) -> types_.ActionRepor
     table_row = types_.TableRow(level=action.level, path=action.path, navlink=action.navlink)
     return types_.ActionReport(
         table_row=table_row,
-        location=_absolute_url(table_row.navlink.link, discourse=discourse),
+        location=_absolute_url(action=action, url=table_row.navlink.link, discourse=discourse),
         result=types_.ActionResult.SUCCESS,
         reason=None,
     )
@@ -115,14 +141,12 @@ class UpdateCase(str, Enum):
     """The possible cases for the update action.
 
     Attrs:
-        INVALID: The action is not valid.
         DRY_RUN: Do not make any changes.
         CONTENT_CHANGE: The content has been changed.
         BASE_MISSING: The base content is not available.
         DEFAULT: No other specific case applies.
     """
 
-    INVALID = "invalid"
     DRY_RUN = "dry-run"
     CONTENT_CHANGE = "content-change"
     BASE_MISSING = "base-missing"
@@ -139,21 +163,18 @@ def _get_update_case(action: types_.UpdateAction, dry_run: bool) -> UpdateCase:
     Returns:
         The named case for the action execution.
     """
-    # Check that action is valid
-    if action.navlink_change.new.link is not None and action.content_change is None:
-        return UpdateCase.INVALID
     if dry_run:
         return UpdateCase.DRY_RUN
     if (
         not dry_run
+        and isinstance(action, types_.UpdatePageAction)
         and action.navlink_change.new.link is not None
-        and action.content_change is not None
         and action.content_change.base is not None
         and action.content_change.local != action.content_change.server
     ):
         return UpdateCase.CONTENT_CHANGE
     if (
-        action.content_change is not None
+        isinstance(action, types_.UpdatePageAction)
         and action.content_change.base is None
         and action.content_change.local != action.content_change.server
     ):
@@ -173,12 +194,9 @@ def _update(
 
     Returns:
         A report on the outcome of executing the action.
-
-    Raises:
-        ActionError: if the content change or new content for a page is None.
     """
     logging.info("dry run: %s, action: %s", dry_run, action)
-    if action.content_change is not None:
+    if isinstance(action, types_.UpdatePageAction):
         _log_content_change(
             base=action.content_change.base or action.content_change.server,
             new=action.content_change.local,
@@ -188,14 +206,11 @@ def _update(
 
     reason: str | None
     match update_case:
-        case UpdateCase.INVALID:
-            raise exceptions.ActionError(
-                f"internal error, content change for page is None, {action=!r}"
-            )
         case UpdateCase.DRY_RUN:
             result = types_.ActionResult.SKIP
             reason = DRY_RUN_REASON
         case UpdateCase.CONTENT_CHANGE:
+            action = typing.cast(types_.UpdatePageAction, action)
             try:
                 content_change = typing.cast(types_.ContentChange, action.content_change)
                 merged_content = content.merge(
@@ -218,7 +233,7 @@ def _update(
             result = types_.ActionResult.SUCCESS
             reason = None
 
-    url = _absolute_url(action.navlink_change.new.link, discourse=discourse)
+    url = _absolute_url(action=action, url=action.navlink_change.new.link, discourse=discourse)
     table_row = types_.TableRow(
         level=action.level, path=action.path, navlink=action.navlink_change.new
     )
@@ -244,19 +259,23 @@ def _delete(
     """
     logging.info("dry run: %s, delete pages: %s, action: %s", dry_run, delete_pages, action)
 
-    url = _absolute_url(action.navlink.link, discourse=discourse)
-    is_group = action.navlink.link is None
+    # Handle group and external references
+    if isinstance(action, (types_.DeleteGroupAction, types_.DeleteExternalRefAction)):
+        return types_.ActionReport(
+            table_row=None,
+            location=None,
+            result=types_.ActionResult.SKIP if dry_run else types_.ActionResult.SUCCESS,
+            reason=DRY_RUN_REASON if dry_run else None,
+        )
+
+    url = _absolute_url(action=action, url=action.navlink.link, discourse=discourse)
     if dry_run:
         return types_.ActionReport(
             table_row=None, location=url, result=types_.ActionResult.SKIP, reason=DRY_RUN_REASON
         )
-    if not delete_pages and not is_group:
+    if not delete_pages and isinstance(action, types_.DeletePageAction):
         return types_.ActionReport(
             table_row=None, location=url, result=types_.ActionResult.SKIP, reason=NOT_DELETE_REASON
-        )
-    if is_group:
-        return types_.ActionReport(
-            table_row=None, location=url, result=types_.ActionResult.SUCCESS, reason=None
         )
 
     try:
@@ -299,18 +318,17 @@ def _run_one(
         ActionError: if an action that is not handled is passed to the function.
     """
     match type(action):
-        case types_.CreateAction:
-            # To help mypy (same for the rest of the asserts), it is ok if the assert does not run
-            assert isinstance(action, types_.CreateAction)  # nosec
+        case types_.CreatePageAction | types_.CreateGroupAction | types_.CreateExternalRefAction:
+            action = typing.cast(types_.CreateAction, action)
             report = _create(action=action, discourse=discourse, dry_run=dry_run, name=name)
-        case types_.NoopAction:
-            assert isinstance(action, types_.NoopAction)  # nosec
+        case types_.NoopPageAction | types_.NoopGroupAction | types_.NoopExternalRefAction:
+            action = typing.cast(types_.NoopAction, action)
             report = _noop(action=action, discourse=discourse)
-        case types_.UpdateAction:
-            assert isinstance(action, types_.UpdateAction)  # nosec
+        case types_.UpdatePageAction | types_.UpdateGroupAction | types_.UpdateExternalRefAction:
+            action = typing.cast(types_.UpdateAction, action)
             report = _update(action=action, discourse=discourse, dry_run=dry_run)
-        case types_.DeleteAction:
-            assert isinstance(action, types_.DeleteAction)  # nosec
+        case types_.DeletePageAction | types_.DeleteGroupAction | types_.DeleteExternalRefAction:
+            action = typing.cast(types_.DeleteAction, action)
             report = _delete(
                 action=action,
                 discourse=discourse,
@@ -362,9 +380,7 @@ def _run_index(
     match type(action):
         case types_.CreateIndexAction:
             try:
-                # To help mypy (same for the rest of the asserts), it is ok if the assert does not
-                # run
-                assert isinstance(action, types_.CreateIndexAction)  # nosec
+                action = typing.cast(types_.CreateIndexAction, action)
                 url = discourse.create_topic(title=action.title, content=action.content)
                 report = types_.ActionReport(
                     table_row=None, location=url, result=types_.ActionResult.SUCCESS, reason=None
@@ -377,7 +393,7 @@ def _run_index(
                     reason=str(exc),
                 )
         case types_.NoopIndexAction:
-            assert isinstance(action, types_.NoopIndexAction)  # nosec
+            action = typing.cast(types_.NoopIndexAction, action)
             report = types_.ActionReport(
                 table_row=None,
                 location=action.url,
@@ -385,8 +401,8 @@ def _run_index(
                 reason=None,
             )
         case types_.UpdateIndexAction:
+            action = typing.cast(types_.UpdateIndexAction, action)
             try:
-                assert isinstance(action, types_.UpdateIndexAction)  # nosec
                 _log_content_change(base=action.content_change.old, new=action.content_change.new)
                 discourse.update_topic(url=action.url, content=action.content_change.new)
                 report = types_.ActionReport(
@@ -396,7 +412,6 @@ def _run_index(
                     reason=None,
                 )
             except exceptions.DiscourseError as exc:
-                assert isinstance(action, types_.UpdateIndexAction)  # nosec
                 report = types_.ActionReport(
                     table_row=None,
                     location=action.url,
@@ -443,7 +458,7 @@ def run_all(
         for action in actions
     ]
     table_rows = (report.table_row for report in action_reports if report.table_row is not None)
-    index_action = reconcile.index_page(index=index, table_rows=table_rows)
+    index_action = reconcile.index_page(index=index, table_rows=table_rows, discourse=discourse)
     index_action_report = _run_index(action=index_action, discourse=discourse, dry_run=dry_run)
     action_reports.append(index_action_report)
     return str(index_action_report.location), action_reports