+++ CHANGELOG.md
@@ -5,6 +5,13 @@
 ### Added
 
 - Support for defining a contents index in the `index.md` file.
+- Support for hidden items on the contentx index.
+- Support for external items on the contents index
+
+### Fixed
+
+- Navigation table not getting removed if the contents was hidden using the
+  `[details=...]...[/details]` tag.
 
 ### Fixed